fix autoloading
