+++ src/DispatchPolicy.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace GwangJinKim\Multidispatch;
+
+class DispatchPolicy {
+    public const FIRST_WINS = 'first-wins';
+    public const LAST_WINS = 'last-wins';
+}
