+++ src/Functions.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace GwangJinKim\Multidispatch;
+
+function multidispatch(): Multidispatch
+{
+    return new Multidispatch();
+}
